#refactoring: (refactor) changed canvas structure, enums file
