+++ packages/web/src/FPClient.ts
@@ -1,9 +1,6 @@
 import { descriptors } from './clientDescriptors/index';
 import type { Descriptors } from './clientDescriptors/index';
-import type {
-  DescriptorsRes,
-  FPClientInterface,
-} from './types/index';
+import type { DescriptorsRes, FPClientInterface } from './types/index';
 import { x64 } from './utils/hashing';
 import { loadClientDescriptors } from './utils/load';
 import { hasProperty } from './utils/utils';
